+++ internal/hdhomerun/guide_test.go
@@ -1,0 +1,30 @@
+package hdhomerun
+
+import (
+	"testing"
+)
+
+func TestAnalyzeGuideXMLStats_sample(t *testing.T) {
+	raw := []byte(`<?xml version="1.0"?>
+<tv>
+  <channel id="1"><display-name>A</display-name></channel>
+  <channel id="2"><display-name>B</display-name></channel>
+  <programme start="20200101000000" channel="1"><title>X</title></programme>
+</tv>`)
+	st, err := AnalyzeGuideXMLStats(raw)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if st.ChannelTags != 2 || st.Programmes != 1 {
+		t.Fatalf("got %+v", st)
+	}
+}
+
+func TestGuideURLFromBase(t *testing.T) {
+	if g := GuideURLFromBase("http://192.168.1.10"); g != "http://192.168.1.10/guide.xml" {
+		t.Fatalf("%q", g)
+	}
+	if g := GuideURLFromBase("http://x/guide.xml"); g != "http://x/guide.xml" {
+		t.Fatalf("%q", g)
+	}
+}